move things out of the logic
